requirements.txt
django-allauth
appdirs==1.4.4
APScheduler==3.9.1.post1
asgiref==3.6.0
astroid==2.4.2
beautifulsoup4==4.9.3
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.4
cryptography==38.0.1
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.1
Django==3.2.5
django-allauth==0.44.0
django-bootstrap4==3.0.1
django-import-export==2.7.1
djangorestframework==3.12.4
et-xmlfile==1.1.0
filelock==3.0.12
gunicorn==20.0.4
idna==3.4
importlib-metadata==1.7.0
isort==5.6.4
lazy-object-proxy==1.4.3
MarkupPy==1.14
mccabe==0.6.1
mysqlclient==2.0.3
oauthlib==3.2.1
odfpy==1.4.1
openpyxl==3.0.10
Pillow==8.2.0
pipenv==2020.8.13
pyactiveresource==2.2.1
pycparser==2.21
PyJWT==2.5.0
pylint==2.6.0
python-dotenv==0.21.1
python3-openid==3.2.0
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
ShopifyAPI==8.0.1
six==1.16.0
soupsieve==2.0.1
sqlparse==0.4.3
tablib==3.0.0
toml==0.10.1
types-cryptography==3.3.23
typing-extensions==4.5.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
virtualenv==20.0.31
virtualenv-clone==0.5.4
wrapt==1.12.1
xlrd==2.0.1
xlwt==1.3.0
zipp==3.2.0
